Move ens domain fetch to user services
