+++ app/src/shared/modules/user/services.js
@@ -7,6 +7,8 @@ import type { Address } from '$shared/flowtype/web3-types'
 import { getDataTokenBalance, getEthBalance } from '$mp/utils/web3'
 import routes from '$routes'
 
+const GRAPH_API_URL = 'https://api.thegraph.com/subgraphs/name/ensdomains/ens'
+
 export const getUserData = (): ApiResult<User> => get({
     url: routes.api.currentUser.index({
         noCache: Date.now(),
@@ -65,3 +67,33 @@ export async function getBalance({ address, type, usePublicNode = false }: GetBa
 
     return balance
 }
+
+type Domains = {
+    data: {
+        domains: Array<{
+            name: string,
+        }>
+    },
+}
+
+export const getEnsDomains = ({ addresses }: {
+    addresses: Array<Address>,
+}): ApiResult<Domains> => post({
+    url: GRAPH_API_URL,
+    data: {
+        query: `
+            query {
+                domains(
+                    where: { owner_in: [${(addresses || []).map((address) => `"${address}"`).join(', ')}]}
+                    orderBy: name
+                ) {
+                    id
+                    name
+                    labelName
+                    labelhash
+                }
+            }
+        `,
+    },
+    useAuthorization: false,
+})
