+++ app/src/userpages/components/StreamPage/New.jsx
@@ -26,7 +26,7 @@ import useModal from '$shared/hooks/useModal'
 import ConfirmDialog from '$shared/components/ConfirmDialog'
 import SvgIcon from '$shared/components/SvgIcon'
 import usePreventNavigatingAway from '$shared/hooks/usePreventNavigatingAway'
-import { getEnsDomains } from '$shared/modules/integrationKey/services'
+import { getEnsDomains } from '$shared/modules/user/services'
 import Spinner from '$shared/components/Spinner'
 import Button from '$shared/components/Button'
 import { truncate } from '$shared/utils/text'
